feat/#510: 아티클 상세 정보 불러오기 api 연결
