+++ app/[team]/dashboard/statistics/[articleId]/page.tsx
@@ -1,0 +1,7 @@
+import StatisticsDetailTemplate from '@/components/statistics/detail/StatisticsDetailTemplate';
+
+const page = () => {
+  return <StatisticsDetailTemplate />;
+};
+
+export default page;
