+++ components/statistics/ArticleList.tsx
@@ -5,13 +5,16 @@ import styled from 'styled-components';
 import { useGetArticleStatisticsList } from '@/hooks/dashboard';
 import { ArticleStatics } from '@/types/dashboard';
 
+import LoadingLottie from '../common/ui/LoadingLottie';
+
 import ArticleField from './ArticleField';
 import ArticleItem from './ArticleItem';
 
 const ArticleList = () => {
   const { team } = useParams();
   const data = useGetArticleStatisticsList(String(team));
-  if (!data) return;
+
+  if (!data) return <LoadingLottie width={4} height={4} />;
 
   const ArticleList = data.data.map((props: ArticleStatics) => {
     return <ArticleItem key={props.id} {...props} />;
